+++ membership-attribute-service/test/controllers/AccountControllerTest.scala
@@ -6,13 +6,15 @@ import org.specs2.mutable.Specification
 import play.api.test.Helpers._
 import play.api.test._
 
+import services.FakePostgresService
+
 class AccountControllerTest extends Specification with Mockito {
 
   "validateContributionAmountUpdateForm" should {
 
     val subName = "s1"
     val commonActions = mock[CommonActions]
-    val controller = new AccountController(commonActions, stubControllerComponents())
+    val controller = new AccountController(commonActions, stubControllerComponents(), FakePostgresService)
     val request = FakeRequest("POST", s"/api/update/amount/contributions/$subName")
 
     "succeed when given value is valid" in {
